requirements.txt
jax==0.4.11
requests
Pillow
torch
transformers
matplotlib
numpy
torchvision
cocotools
transformers[torch]
accelerate
cython
datasets